LayoutTests/http/tests/inspector/filesystem/get-filesystem-root-expected.txt

Tests getFileSystemRoot.

PASS: filesystem:http://127.0.0.1:8000/temporary/ is found.
